+++ polars_bio/__init__.py
@@ -50,6 +50,7 @@ if "POLARS_FORCE_NEW_STREAMING" not in os.environ:
 register_gff = data_processing.register_gff
 register_gtf = data_processing.register_gtf
 register_vcf = data_processing.register_vcf
+register_bcf = data_processing.register_bcf
 register_vcf_zarr = data_processing.register_vcf_zarr
 register_fastq = data_processing.register_fastq
 register_fasta = data_processing.register_fasta
@@ -65,6 +66,7 @@ register_view = data_processing.register_view
 sql = data_processing.sql
 
 describe_vcf = data_input.describe_vcf
+describe_bcf = data_input.describe_bcf
 describe_vcf_zarr = data_input.describe_vcf_zarr
 describe_bam = data_input.describe_bam
 describe_sam = data_input.describe_sam
@@ -77,6 +79,7 @@ read_fastq = data_input.read_fastq
 read_gff = data_input.read_gff
 read_gtf = data_input.read_gtf
 read_table = data_input.read_table
+read_bcf = data_input.read_bcf
 read_vcf = data_input.read_vcf
 read_vcf_zarr = data_input.read_vcf_zarr
 read_fastq = data_input.read_fastq
@@ -97,6 +100,7 @@ scan_gff = data_input.scan_gff
 scan_gtf = data_input.scan_gtf
 scan_pairs = data_input.scan_pairs
 scan_table = data_input.scan_table
+scan_bcf = data_input.scan_bcf
 scan_vcf = data_input.scan_vcf
 scan_vcf_zarr = data_input.scan_vcf_zarr
 write_vcf = data_input.write_vcf
@@ -157,6 +161,7 @@ __all__ = [
     "print_metadata_json",
     "print_metadata_summary",
     # I/O functions
+    "describe_bcf",
     "describe_vcf",
     "describe_vcf_zarr",
     "describe_bam",
@@ -175,6 +180,7 @@ __all__ = [
     "read_gff",
     "read_gtf",
     "read_table",
+    "read_bcf",
     "read_vcf",
     "read_vcf_zarr",
     "scan_bam",
@@ -189,6 +195,7 @@ __all__ = [
     "scan_gtf",
     "scan_pairs",
     "scan_table",
+    "scan_bcf",
     "scan_vcf",
     "scan_vcf_zarr",
     "write_vcf",
@@ -205,6 +212,7 @@ __all__ = [
     "sink_cram",
     "register_gff",
     "register_gtf",
+    "register_bcf",
     "register_vcf",
     "register_vcf_zarr",
     "register_fastq",